refactor(website): extract buildPanes for reuse by library pages
